+++ Graphics/Implicit/Definitions.hs
@@ -61,6 +61,8 @@ module Graphics.Implicit.Definitions (
         Transform3,
         Torus,
         Ellipsoid,
+        BoxFrame,
+        Link,
         Extrude,
         ExtrudeM,
         ExtrudeOnEdgeOf,
@@ -332,6 +334,8 @@ data SymbolicObj3 =
     | Cylinder ℝ ℝ ℝ --
     | Torus ℝ ℝ
     | Ellipsoid ℝ ℝ ℝ
+    | BoxFrame ℝ3 ℝ -- b e from https://iquilezles.org/articles/distfunctions/
+    | Link ℝ ℝ ℝ -- le r1 r2 from https://iquilezles.org/articles/distfunctions/
     -- Simple transforms
     | Rotate3 (Quaternion ℝ) SymbolicObj3
     | Transform3 (M44 ℝ) SymbolicObj3
@@ -359,6 +363,8 @@ instance Show SymbolicObj3 where
     -- centered.
     Cube sz -> showCon "cube" @| False @| sz
     Sphere d -> showCon "sphere" @| d
+    BoxFrame b e -> showCon "boxFrame" @| b @| e
+    Link le r1 r2 -> showCon "link" @| le @| r1 @| r2
     -- NB: The arguments to 'Cylinder' are backwards compared to 'cylinder' and
     -- 'cylinder2'.
     Cylinder h r1 r2 | r1 == r2 ->